getting default styling code for arrow component
