+++ components/LessonSection/TeachIt/index.js
@@ -18,7 +18,8 @@ import Link from "next/link";
 import Button from "../../General/Button";
 import { getIsValObj, getObjVals } from "../../../globalFns";
 import { UNVIEWABLE_LESSON_STR } from "../../../globalVars";
-import ClickMeArrow from "../../ClickMeArrow";
+import Arrow from "../../ClickMeArrow";
+import throttle from "lodash.throttle";
 
 const LessonTile = ({
   lessonTileUrl,
@@ -124,6 +125,20 @@ const TeachIt = ({
     setSelectedGradeResources(selectedGrade.links);
     setSelectedGrade(selectedGrade);
   };
+
+  let timer;
+
+  const handleElementVisibility = throttle(() => {
+    clearTimeout(timer);
+
+    if (arrowContainer.inViewport) {
+      setArrowContainer(state => ({ ...state, isInView: true }));
+
+      timer = setTimeout(() => {
+        setArrowContainer(state => ({ ...state, isInView: false }));
+      }, 3500);
+    }
+  }, 100);
 
   useEffect(() => {
     const lessonPartPath = window.location.href.split("#").at(-1);
@@ -323,7 +338,7 @@ const TeachIt = ({
                 {...lessonTilesObj}
                 removeClickToSeeMoreTxt={removeClickToSeeMoreTxt}
                 key={`${index}_part`}
-                ClickToSeeMoreComp={index === 0 ? <ClickMeArrow _arrowContainer={[arrowContainer, setArrowContainer]} /> : null}
+                ClickToSeeMoreComp={index === 0 ? <Arrow handleElementVisibility={handleElementVisibility} /> : null}
                 FeedbackComp={(part.status === "Beta") ? (
                   <SendFeedback
                     parentDivStyles={{ backgroundColor: '#EBD0FF', zIndex: 100, border: '1px solid #B7B6C2' }}
